[test] split onnx export test out of test_jit
as titled, to reduce size of test_jit

[ghstack-poisoned]
